Created dialog_utils.dart to reduce code duplication
